Some refactoring
Separating the YAML loading and the validation in the Parser class.
