+++ org.eclipse.winery.lsp/src/main/java/org/eclipse/winery/lsp/Client/VSCodeExtention/toscalsp/src/extension.ts
@@ -14,7 +14,7 @@ export function activate(context: vscode.ExtensionContext) {
         logChannel.appendLine(`Java Executable: ${executable}`);
         logChannel.appendLine(`Classpath: ${classPath}`);
         logChannel.appendLine(`Args: ${args.join(' ')}`);
-        
+
         const serverOptions: ServerOptions = {
             command: executable,
             args,
